Add dqn with candle (#1)
WIP: Add dqn with candle (#1)

WIP: Add dqn with candle (#1)

Add dqn with candle (#1)

+++ border-candle-agent/src/mlp/config.rs
@@ -4,9 +4,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize, Serialize, PartialEq, Clone)]
 /// Configuration of [`Mlp`](super::Mlp).
 pub struct MlpConfig {
-    pub(super) in_dim: usize,
-    pub(super) units: Vec<usize>,
-    pub(super) out_dim: usize,
+    pub(super) in_dim: i64,
+    pub(super) units: Vec<i64>,
+    pub(super) out_dim: i64,
     pub(super) activation_out: bool,
 }
 
@@ -14,7 +14,7 @@ impl MlpConfig {
     /// Creates configuration of MLP.
     ///
     /// * `activation_out` - If `true`, activation function is added in the final layer.
-    pub fn new(in_dim: usize, units: Vec<usize>, out_dim: usize, activation_out: bool) -> Self {
+    pub fn new(in_dim: i64, units: Vec<i64>, out_dim: i64, activation_out: bool) -> Self {
         Self {
             in_dim,
             units,
@@ -25,11 +25,11 @@ impl MlpConfig {
 }
 
 impl OutDim for MlpConfig {
-    fn get_out_dim(&self) -> usize {
+    fn get_out_dim(&self) -> i64 {
         self.out_dim
     }
 
-    fn set_out_dim(&mut self, out_dim: usize) {
+    fn set_out_dim(&mut self, out_dim: i64) {
         self.out_dim = out_dim;
     }
 }
